Fixed failing async_io_lstat_works unit test under standalone.
Fixed failing async_io_errors_test unit test under standalone (it was actually caused by a slow CPU).

Made BOOST_AFIO_USE_LEGACY_FILESYSTEM_SEMANTICS to retain compatibility with Boost.Filesystem being outdated relative to the Filesystem TS.

+++ include/boost/afio/detail/Utility.hpp
@@ -50,7 +50,7 @@ BOOST_AFIO_V1_NAMESPACE_BEGIN
                 THREADNAME_INFO info;
                 info.dwType = 0x1000;
                 info.szName = threadName;
-                info.dwThreadID = -1;
+                info.dwThreadID = (DWORD) -1;
                 info.dwFlags = 0;
 
                 __try
